requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[argon2]==1.7.4
cryptography==42.0.1
Pillow==10.2.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
rawpy>=0.19.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
numpy<2.0.0
opencv-python-headless
requests